+++ tests/packerlicious/test_builder_virtualbox.py
@@ -22,7 +22,7 @@ class TestVirtualBoxIsoBuilder(object):
             b.to_dict()
         assert 'required' in str(excinfo.value)
 
-    def test_is_checksum_mutually_exclusive(self):
+    def test_iso_checksum_mutually_exclusive(self):
         b = builder.VirtualboxIso(
             iso_url="/url/to/iso",
             iso_checksum_type=builder.VirtualboxIso.MD5,